requirements.txt
fastapi
uvicorn
pymongo
python-multipart
python-dotenv

insightface==0.7.3
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
imutils
boto3